server/requirements.txt
annotated-doc 
annotated-types 
anyio 
certifi 
click 
colorama 
fastapi 
groq 
h11 
httpcore 
httptools 
httpx 
idna 
iniconfig 
packaging 
pluggy 
pydantic 
pydantic_core 
Pygments 
pytest 
python-multipart 
python-dotenv 
PyYAML 
starlette 
typing-inspection 
typing_extensions 
uvicorn 
watchfiles 
websockets 
